ln-simln-jamming: move sink interceptor to dedicated file
